@@ -20,7 +20,7 @@
     "@angular/platform-browser": "^17.0.9",
     "@angular/platform-browser-dynamic": "^17.0.9",
     "@angular/router": "^17.0.9",
-    "@auth0/auth0-angular": "^2.2.1",
+    "angular-oauth2-oidc": "^17.0.1",
     "ngx-mat-multi-sort": "^17.0.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
@@ -38,6 +38,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.3.3"
   }
 }
