@@ -8,7 +8,7 @@ import {MatIconModule} from "@angular/material/icon";
 import {MatToolbarModule} from "@angular/material/toolbar";
 import {AboutComponent} from './components/about/about.component';
 import {CaseExplorerComponent} from './components/case-explorer/case-explorer.component';
-import {HTTP_INTERCEPTORS, HttpClientModule} from "@angular/common/http";
+import {HttpClientModule} from "@angular/common/http";
 import {MatSortModule} from "@angular/material/sort";
 import {RegistryViewerComponent} from './components/registry-viewer/registry-viewer.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
@@ -45,13 +45,14 @@ import {ConfigService} from "./service/config.service";
 import {MatSnackBarModule} from "@angular/material/snack-bar";
 import {MatDatepickerModule} from "@angular/material/datepicker";
 import {MatTooltipModule} from "@angular/material/tooltip";
-import {APP_DATE_FORMATS, AppDateAdapter} from "./provider/format-datepicker";
-import {AuthHttpInterceptor, AuthModule} from '@auth0/auth0-angular';
-import { Auth0LoginComponent } from './auth0-login/auth0-login.component';
+import {APP_DATE_FORMATS, AppDateAdapter} from "./provider/format-datepicker"
 import {MatMenuModule} from "@angular/material/menu";
 import {MatDividerModule} from "@angular/material/divider";
 import { LandingComponent } from './components/landing/landing.component';
 import {MatRadioModule} from "@angular/material/radio";
+import {OAuthModule} from "angular-oauth2-oidc";
+import {authModuleConfig} from "../assets/config/auth-module-config";
+import {LoginComponent} from "./components/login/login.component";
 
 export const configFactory = (configService: ConfigService) => {
   return () => configService.loadConfig();
@@ -76,8 +77,8 @@ export const configFactory = (configService: ConfigService) => {
     SectionComponent,
     ConformationDialogComponent,
     DemoModeComponent,
-    Auth0LoginComponent,
     LandingComponent,
+    LoginComponent
   ],
   imports: [
     BrowserModule,
@@ -110,31 +111,12 @@ export const configFactory = (configService: ConfigService) => {
     MatDatepickerModule,
     NgOptimizedImage,
     MatTooltipModule,
-    AuthModule.forRoot({
-        domain: 'grady-temp.us.auth0.com',
-        clientId: '4T568aTy0dj7keOCla7FubFQO7hJ9iiH',
-        authorizationParams: {
-          redirect_uri: window.location.href,
-          audience: 'http://smartchartsuite.grady/registry-viewer-api/',
-          scope: 'profile email openid read:scd read:syphilis write:metadata write:scd write:syphilis'
-        },
-        httpInterceptor: {
-          allowedList: [{
-            uri: 'https://smartchartsuite.dev.heat.icl.gtri.org/registry-viewer-api/*',
-            // tokenOptions: {
-            //   authorizationParams: {
-            //     audience: 'http://smartchartsuite.grady/registry-viewer-api/'
-            //   }
-            // }
-          }
-
-          ]
-        }
-      }
+    OAuthModule.forRoot(
+      authModuleConfig
     ),
     MatMenuModule,
     MatDividerModule,
-    MatRadioModule
+    MatRadioModule,
   ],
   providers: [
     DrawerService,
@@ -147,7 +129,6 @@ export const configFactory = (configService: ConfigService) => {
       deps: [ConfigService],
       multi: true
     },
-    { provide: HTTP_INTERCEPTORS, useClass: AuthHttpInterceptor, multi: true },
   ],
   bootstrap: [AppComponent]
 })
