@@ -3,15 +3,14 @@ import { RouterModule, Routes } from '@angular/router';
 import {AboutComponent} from "./components/about/about.component";
 import {CaseExplorerComponent} from "./components/case-explorer/case-explorer.component";
 import {RegistryViewerComponent} from "./components/registry-viewer/registry-viewer.component";
-import { AuthGuard } from '@auth0/auth0-angular';
 import {LandingComponent} from "./components/landing/landing.component";
+import {AuthGuard} from "./guards/auth.guard";
 
 const routes: Routes = [
   //We need to add Login guard to this section
   {
     path: '',
     component: LandingComponent,
-    canActivate: [AuthGuard],
   },
   {
     path: 'case',
